move tagSelector to common handlers
